Move texture data into generators common
