+++ src/generators/_common/textureData.test.ts
