+++ src/generators/minecraftItem/ui/textureVersions.ts
@@ -3,22 +3,22 @@ import {
   type TextureData,
   type TextureFrame,
   tilesToTextureFrames,
-} from "@genroot/builder/modules/textureData";
+} from "../../_common/textureData";
 
-import * as Texture_1_7_10_Items from "@genroot/textures/texture_minecraft_1_7_10_items";
-import * as Texture_1_7_10_Blocks from "@genroot/textures/texture_minecraft_1_7_10_blocks";
+import * as Texture_1_7_10_Items from "../../../textures/texture_minecraft_1_7_10_items";
+import * as Texture_1_7_10_Blocks from "../../../textures/texture_minecraft_1_7_10_blocks";
 
-import * as Texture_1_13_2_Items from "@genroot/textures/texture_minecraft_1_13_2_items";
-import * as Texture_1_13_2_Blocks from "@genroot/textures/texture_minecraft_1_13_2_blocks";
+import * as Texture_1_13_2_Items from "../../../textures/texture_minecraft_1_13_2_items";
+import * as Texture_1_13_2_Blocks from "../../../textures/texture_minecraft_1_13_2_blocks";
 
-import * as Texture_1_18_2_Items from "@genroot/textures/texture_minecraft_1_18_2_items";
-import * as Texture_1_18_2_Blocks from "@genroot/textures/texture_minecraft_1_18_2_blocks";
+import * as Texture_1_18_2_Items from "../../../textures/texture_minecraft_1_18_2_items";
+import * as Texture_1_18_2_Blocks from "../../../textures/texture_minecraft_1_18_2_blocks";
 
-import * as Texture_1_20_4_Items from "@genroot/textures/texture_minecraft_1_20_4_items";
-import * as Texture_1_20_4_Blocks from "@genroot/textures/texture_minecraft_1_20_4_blocks";
+import * as Texture_1_20_4_Items from "../../../textures/texture_minecraft_1_20_4_items";
+import * as Texture_1_20_4_Blocks from "../../../textures/texture_minecraft_1_20_4_blocks";
 
-import * as Texture_26_1_2_Items from "@genroot/textures/texture_minecraft_26_1_2_items";
-import * as Texture_26_1_2_Blocks from "@genroot/textures/texture_minecraft_26_1_2_blocks";
+import * as Texture_26_1_2_Items from "../../../textures/texture_minecraft_26_1_2_items";
+import * as Texture_26_1_2_Blocks from "../../../textures/texture_minecraft_26_1_2_blocks";
 
 const definitions: [TextureData, number][] = [
   [Texture_1_7_10_Items.data, 16],
